Added options for cache directory for faster writing to network dirves and option whether to replace files.

+++ compress_tiffs.py
@@ -1,11 +1,12 @@
 import os
 import argparse
 import shutil
+from time import sleep
 import tifffile
 from tqdm import tqdm
 
 
-def compress_tiff_files(input_path, quality, compression, threads):
+def compress_tiff_files(input_path, quality, compression, threads, cache_dir, replace_files):
     """
     Compresses the TIFF files in the given input path (folder or file) using the specified compression algorithm and quality percentage.
     Creates a new compressed TIFF file with the name '.part' and replaces the original file(s).
@@ -21,13 +22,22 @@ def compress_tiff_files(input_path, quality, compression, threads):
         with tqdm(total=len(tiff_files), ncols=80, desc="Progress") as pbar:
             for file_path in tiff_files:
                 temp_file_path = file_path + '.part'
+                temp_file_path_beside_original = temp_file_path
                 try:
+                    if cache_dir:
+                        cache_file_path = os.path.join(cache_dir, os.path.basename(temp_file_path))
+                        temp_file_path = cache_file_path
+
                     # Compress the TIFF file using the specified algorithm and quality
                     with tifffile.TiffWriter(temp_file_path) as tiff:
                         tiff.write(tifffile.imread(file_path), compression=(compression, quality), maxworkers=threads)
 
-                    # Replace the original file with the compressed file
-                    shutil.move(temp_file_path, file_path)
+                    if cache_dir:
+                        # Move the compressed file from the temporary cache directory to the final destination
+                        shutil.move(temp_file_path, temp_file_path_beside_original)
+                    if replace_files:
+                        # Replace the original file with the compressed file
+                        shutil.move(temp_file_path_beside_original, file_path)
                     pbar.update(1)
                     print(f"Compressed: {file_path}")
                 except Exception as e:
@@ -37,13 +47,24 @@ def compress_tiff_files(input_path, quality, compression, threads):
         # Input path is a file
         file_path = input_path
         temp_file_path = file_path + '.part'
+        temp_file_path_beside_original = temp_file_path
+        
+
         try:
+            if cache_dir:
+                cache_file_path = os.path.join(cache_dir, os.path.basename(temp_file_path))
+                temp_file_path = cache_file_path
             # Compress the TIFF file using the specified algorithm and quality
             with tifffile.TiffWriter(temp_file_path) as tiff:
                 tiff.write(tifffile.imread(file_path), compression=(compression, quality), maxworkers=threads)
 
-            # Replace the original file with the compressed file
-            shutil.move(temp_file_path, file_path)
+            if cache_dir:
+                # Move the compressed file from the temporary cache directory to the final destination
+                shutil.move(temp_file_path, temp_file_path_beside_original)
+                
+            if replace_files:
+                # Replace the original file with the compressed file
+                shutil.move(temp_file_path, file_path)
             print(f"Compressed: {file_path}")
         except Exception as e:
             print(f"Error compressing: {file_path}")
@@ -73,10 +94,20 @@ if __name__ == '__main__':
                         type=int,
                         help="Maximum number of threads to use (should be less that CPU cores).", 
                         default=None)
+    parser.add_argument(
+                        '--cache_dir', 
+                        type=str,
+                        help="Temporary directory on a fast local SSD, for better writing performance to network drives.", 
+                        default=None)
+    parser.add_argument(
+                        '--do_not_replace', 
+                        action="store_true",
+                        help="Replace files at the destination? If enabled places file with extension '.part' beside original.", 
+                        default=False)
 
     args = parser.parse_args()
 
     if args.folder:
-        compress_tiff_files(args.folder, args.quality, args.compression, args.threads)
+        compress_tiff_files(args.folder, args.quality, args.compression, args.threads, args.cache_dir, not args.do_not_replace)
     elif args.file:
-        compress_tiff_files(args.file, args.quality, args.compression, args.threads)
+        compress_tiff_files(args.file, args.quality, args.compression, args.threads, args.cache_dir, not args.do_not_replace)
